+++ backend/src/pageHandlers/mainHandler.js
@@ -1,0 +1,24 @@
+import * as dbHandler from "../db/dbHandler.js"
+
+export const GetShelfs = async (req, res) => {
+    const shelfsResult = await dbHandler.GetAllShelfs();
+
+    if (!shelfsResult.success){
+        return res.status(404).json({
+            success: false,
+            message: shelfsResult.reason}
+        );
+    }
+
+    if (shelfsResult.value.rows.length == 0){
+        return res.status(404).json({
+            success: false,
+            message: "no shelfs exist"}
+        );
+    }
+
+    return res.status(200).json({success: true,
+        role: usernameResult.value.rows
+    });
+
+}